+++ front/src/pages/samples/singleCreate/Default.tsx
